add project based query for merged reviews (#33)
Although it's not often used in gerrit. It is possible
to query a configured project.  A gerrit project is a set
of gerrit repos with a given name, this is a server side
config.

In my particular use case it is very helpful to see what
reviews have merged in $x days for the "tripleo" project